+++ src/utils/logging.jl
@@ -30,13 +30,14 @@ const _default_message_loglevel = LogLevel(0)
 const _default_logger = @static if VERSION >= v"1.7.0"
     Ref{Logging.ConsoleLogger}(
         Logging.ConsoleLogger(
+            stdout,
             Logging.Info,
             show_limited=false,
             meta_formatter=meta_formatter_groebner
         )
     )
 else
-    Ref{Logging.ConsoleLogger}(Logging.ConsoleLogger())
+    Ref{Logging.ConsoleLogger}(Logging.ConsoleLogger(stdout))
 end
 
 function meta_formatter_groebner(level::LogLevel, _module, group, id, file, line)