+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Groebner v0.4.5 Release notes
 
+New features
+- Added keyword argument `statistics` to all functions in the interface. It allows one to collect and print some timings and statistics after the computation finishes.
+- Added section `For developers` on the documentation page.
+
 A couple of bugfixes.
 
 Versions of dependencies: